refactor: BaseModal, ContentModal 모달 폴더 이동
